(r) Merge error fix (FAB)

+++ client/src/app/games-card-view/games-card-view.component.ts
@@ -1,4 +1,53 @@
+import { Component, Input, OnInit } from "@angular/core";
+import { Message } from "../../../../common/communication/message";
+import { ICommonGameCard } from "../../../../common/model/gameCard";
+import { GamesCardService } from "../services/games-card.service";
+import { StringFormater } from "../util/stringFormater";
 
+@Component({
+  selector: "app-games-card-view",
+  templateUrl: "./games-card-view.component.html",
+  styleUrls: ["./games-card-view.component.css"],
+})
+export class GamesCardViewComponent implements OnInit {
+  @Input() public gameCard: ICommonGameCard;
+  @Input() public isInAdminView: boolean = false;
+
+  public leftButton: string = "Play";
+  public rightButton: string = "Create";
+  public originalImage: File;
+
+  private gamesCardService: GamesCardService;
+
+  public constructor(gamesCardService: GamesCardService) {
+    this.gamesCardService = gamesCardService;
+  }
+
+  public ngOnInit(): void {
+    if (this.isInAdminView) {
+      this.leftButton = "Delete";
+      this.rightButton = "Reset";
+    }
+  }
+
+  public toMinutes(index: number, times: number[]): string {
+    return StringFormater.secondsToMinutes(times[index]);
+  }
+
+  public onLeftButtonClick(): void {
+    if (this.isInAdminView) {
+      this.deleteGameCard();
+    }
+  }
+
+  public onRightButtonClick(): void {
+    if (this.isInAdminView) {
+      this.resetBestTimes();
+    }
+  }
+
+  public deleteGameCard(): void {
+    if (confirm("Are you sure you want to delete the Game Card called " + this.gameCard.title + "?")) {
       this.gamesCardService.deleteGameCard(this.gameCard.id).subscribe((message: Message) => {
         window.location.reload();
       });
